requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
networkx
reportlab
python-dotenv
httpx